add blogController Trait to similary code in blogController and articleController
