+++ src/Controller/ArticleController.php
@@ -2,8 +2,9 @@
 
 namespace App\Controller;
 
-use App\Repository\{ArticleRepository, CommentRepository};
-use Core\DefaultAbstract\{DefaultAbstractEntity, LoggedAbstractController};
+use App\Repository\ArticleRepository;
+use Core\Traits\Controller\blogControllerTrait;
+use Core\DefaultAbstract\LoggedAbstractController;
 use Core\Exception\CoreException;
 use Core\Traits\Controller\SeeControllerTrait;
 use Exception;
@@ -14,9 +15,8 @@ use Exception;
  */
 class ArticleController extends LoggedAbstractController
 {
-    use SeeControllerTrait;
-
-    public static $entityLabel = "article";
+    use SeeControllerTrait,
+        blogControllerTrait;
 
     /**
      * Action by default
@@ -35,42 +35,6 @@ class ArticleController extends LoggedAbstractController
                 'page'     => 'article'
             ]
         );
-    }
-
-    /**
-     * Give params to seeAction
-     * @return array|mixed[]
-     * @throws Exception
-     */
-    public function getSeeParam(): array
-    {
-        return [
-            'articleId',
-            'article',
-            new ArticleRepository(),
-            'article/articleById.html.twig',
-        ];
-    }
-
-    /**
-     * @param array|mixed[]         $data
-     * @param DefaultAbstractEntity $entity
-     *
-     * @return array|mixed[]
-     */
-    public function preRenderView(array $data, DefaultAbstractEntity $entity): array
-    {
-        // Load comments associate to the articleId
-        $comments         = (new CommentRepository())->search(
-            [
-            'approved' => 'oui',
-            'articleId' => $entity->getId()
-            ]
-        );
-
-        $data['comments'] = $comments;
-
-        return $data;
     }
 
     /**
