+++ src/Controller/BlogController.php
@@ -2,9 +2,9 @@
 
 namespace App\Controller;
 
-use App\Repository\{ArticleRepository, CommentRepository};
-use Core\DefaultAbstract\{DefaultAbstractController, DefaultAbstractEntity};
-use Core\Traits\Controller\{SeeControllerTrait};
+use App\Repository\ArticleRepository;
+use Core\DefaultAbstract\DefaultAbstractController;
+use Core\Traits\Controller\{blogControllerTrait, SeeControllerTrait};
 use Exception;
 
 /**
@@ -13,9 +13,8 @@ use Exception;
  */
 class BlogController extends DefaultAbstractController
 {
-    use SeeControllerTrait;
-
-    public static $entityLabel = "article";
+    use SeeControllerTrait,
+        blogControllerTrait;
 
     /**
      * Action by default
@@ -38,40 +37,5 @@ class BlogController extends DefaultAbstractController
                 'page'     => 'blog'
             ]
         );
-    }
-
-    /**
-     * Give params to seeAction
-     * @return array|mixed[]
-     * @throws Exception
-     */
-    public function getSeeParam(): array
-    {
-        return [
-            'articleId',
-            'article',
-            new ArticleRepository(),
-            'article/articleById.html.twig',
-        ];
-    }
-
-    /**
-     * @param array|mixed[]         $data
-     * @param DefaultAbstractEntity $entity
-     *
-     * @return array|mixed[]
-     */
-    public function preRenderView(array $data, DefaultAbstractEntity $entity): array
-    {
-        $comments         = (new CommentRepository())->search(
-            [
-                'approved' => 'oui',
-                'articleId' => $entity->getId()
-            ]
-        );
-
-        $data['comments'] = $comments;
-
-        return $data;
     }
 }
